reflex_app/backup_20251212_before_restore/requirements.txt
# Core Reflex
reflex==0.8.19

# Database
libsql-client==0.3.1
sqlmodel==0.0.27

# Data processing
pandas==2.2.3
numpy==2.3.4
plotly==5.24.1

# Web/HTTP
httpx==0.28.1
python-dotenv==1.0.1
requests>=2.31.0

# Auth
bcrypt==4.3.0

# Visualization
altair==5.5.0
